% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_points.R
\name{get_points}
\alias{get_points}
\title{Get point value of shot based on distance and angle}
\usage{
get_points(x, y, score)
}
\arguments{
\item{x}{Horizontal distance of the shot taken from the baseline in feet}

\item{y}{Vertical distance of the shot taken from the baseline in feet}

\item{score}{MADE or MISSED}
}
\value{
Point value of shot
}
\description{
Get point value of shot based on distance and angle
}
\examples{
get_points(0, 0) # 2
get_points(23.75, 0) # 3
get_points(20.75, 23) # 2
get_points(22, 100) # 2
get_points(22, 0) # 3
get_points(23.75, 0) # 3
get_points(23.75, 100) # 3
get_points(23.75, 140) # 3

}
