% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shot_distance.R
\name{get_shot_distance}
\alias{get_shot_distance}
\title{Title}
\usage{
get_shot_distance(x, y)
}
\arguments{
\item{x}{Horizontal distance of the shot taken from the baseline in feet}

\item{y}{Vertical distance of the shot taken from the baseline in feet}
}
\value{
The distance of the shot from the basket in feet.
}
\description{
Title
}
\examples{
get_shot_distance(0, 0) # 0
get_shot_distance(0, 10) # 10
get_shot_distance(10, 0) # 10
get_shot_distance(10, 10) # 14.1421356237
}
