% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_three_point.R
\name{is_three_point}
\alias{is_three_point}
\title{Is three point shot based on distance and angle}
\usage{
is_three_point(x, y)
}
\arguments{
\item{x}{Horizontal distance of the shot taken from the baseline in feet}

\item{y}{Vertical distance of the shot taken from the baseline in feet}
}
\value{
TRUE if 3 point shot, FALSE otherwise
}
\description{
Is three point shot based on distance and angle
}
\examples{
get_points(0, 0) # FALSE
get_points(23.75, 0) # TRUE
get_points(20.75, 23) # FALSE
get_points(22, 100) # FALSE
get_points(22, 0) # TRUE
get_points(23.75, 0) # TRUE
get_points(23.75, 100) # TRUE
get_points(23.75, 140) # TRUE

}
