% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_shooting_percentage.R
\name{true_shooting_percentage}
\alias{true_shooting_percentage}
\title{True shooting percentage}
\usage{
true_shooting_percentage(pts, fga, fta, fta_coef = 0.44)
}
\arguments{
\item{pts}{Points scored}

\item{fga}{Field goals attempted}

\item{fta}{Free throws attempted}

\item{fta_coef}{Coefficient used to multiply free throws attempted. Some
formulas use 0.475 as the coefficient. However, the default value is 0.44.
The .44 multiplier is because not all free throws take up a possession.
Technical foul shots and "and-ones" do not, while there are more than two
free throws on one possession with a three-shot foul.
Research has determined that about 44% of all free throws take up
possessions, thus .44 is used as the multiplier.
From https://www.nba.com/thunder/news/stats101.html.}
}
\value{
True shooting percentage
}
\description{
True shooting percentage
}
